Restrictions to employees role

+++ ByL_Client_Management/customer/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, redirect, get_object_or_404
 from django.db.models import Q, Func
 from django.core.paginator import Paginator
 from django.contrib.auth.decorators import login_required
+from django.http import HttpResponseForbidden
 from .models import Customer
 from .forms import CustomerForm
 
@@ -48,16 +49,19 @@ def create_customer(request):
 def edit_customer(request, customer_id):
     customer = get_object_or_404(Customer, id=customer_id)
     if request.method == 'POST':
-        form = CustomerForm(request.POST, instance=customer)
+        form = CustomerForm(request.POST, instance=customer, user=request.user)
         if form.is_valid():
             form.save()
             return redirect('customer_list')
     else:
-        form = CustomerForm(instance=customer)
+        form = CustomerForm(instance=customer, user=request.user)
     return render(request, 'customers/edit_customer.html', {'form': form, 'customer': customer})
 
 @login_required
 def delete_customer(request, customer_id):
+    if request.user.role != 'admin':
+        return HttpResponseForbidden("You do not have permission to delete customers.")
+    
     customer = get_object_or_404(Customer, id=customer_id)
     if request.method == 'POST':
         customer.delete()
